AddArchitectureMVVM
