+++ MyBook App/MyBook_AppApp.swift
@@ -7,12 +7,6 @@
 
 import SwiftUI
 
-// MARK: - Tabs Enum
-enum Tabs: Hashable {
-    case buscar
-    case lista
-}
-
 @main
 struct MyBook_AppApp: App {
     var body: some Scene {
@@ -21,3 +15,4 @@ struct MyBook_AppApp: App {
         }
     }
 }
+
