+++ MyBook App/Person/PersonModel.swift
@@ -1,0 +1,7 @@
+//
+//  PersonModel.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import Foundation
