+++ MyBook App/Person/PersonVM.swift
@@ -1,0 +1,7 @@
+//
+//  PersonVM.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import SwiftUI
