+++ MyBook App/Person/PersonView.swift
@@ -1,0 +1,18 @@
+//
+//  PersonView.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import SwiftUI
+
+struct PersonView: View {
+    var body: some View {
+        ZStack {
+            AnimatedCirclesBackground()
+            
+            Text("PersonView")
+        }
+        
+    }
+}
