+++ MyBook App/Search/SearchModel.swift
@@ -1,0 +1,7 @@
+//
+//  SearchModel.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import Foundation
