+++ MyBook App/Timer/TimerModel.swift
@@ -1,0 +1,7 @@
+//
+//  WorldsModel.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import Foundation
