+++ MyBook App/Timer/TimerVM.swift
@@ -1,0 +1,7 @@
+//
+//  WorldVM.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import SwiftUI
