+++ MyBook App/Timer/TimerView.swift
@@ -1,0 +1,16 @@
+//
+//  WorldsView.swift
+//  MyBook App
+//
+//  Created by ThiagoMotaMachado on 06/09/25.
+//
+import SwiftUI
+
+struct worldView: View {
+    var body: some View {
+        ZStack {
+            AnimatedCirclesBackground()
+            Text("worldView")
+        }
+    }
+}
